After Coding end

+++ UFL-Workflow/branches/request_dynamic_fields/lib/UFL/Workflow/Controller/Processes.pm
@@ -232,7 +232,8 @@ sub add_request : PathPart Chained('process') Args(0) {
 
     if ($c->req->method eq 'POST') {
         my $result = $self->validate_form($c);
-        if ($result->success) {
+	my $result_field = $self->validate_fields($c);
+        if ($result->success and $result_field->success ) {
             my $group = $c->model('DBIC::Group')->find($result->valid('group_id'));
             $c->detach('/default') unless $group;
             #TBD save the fields too after validating.
@@ -251,6 +252,9 @@ sub add_request : PathPart Chained('process') Args(0) {
                 # Make sure we get insert_time and update_time
                 $request->discard_changes;
 
+                if ( my %field_datas = $self->get_field_data($c, $result_field)) {
+		    $request->add_field_data(%field_datas);
+		}
                 if (my $upload = $c->req->upload('document')) {
                     my $document = $request->add_document(
                         $c->user->obj,
@@ -275,10 +279,95 @@ sub add_request : PathPart Chained('process') Args(0) {
     $c->stash(
         process  => $process,
         groups   => $groups,
+	field    => $process->first_field,
         template => 'processes/add_request.tt',
     );
 }
 
+=head2 validate_fields 
+
+Validates the extra fields of this process
+
+=cut
+sub get_field_data {
+    my ($self, $c, $result) = @_;
+
+    my $process = $c->stash->{process};
+    my %data;
+    my $field = $process->first_field;
+    
+    while ($field) {
+        $data{$field->id} = $result->valid($field->id);
+	$field = $field->next_field;
+    }
+
+    return %data;
+}
+=head2 validate_fields 
+
+Validates the extra fields of this process
+
+=cut
+sub validate_fields {
+    my ($self, $c) = @_;
+    
+    # 1. form a yml query based on database.
+    # 2. form a yml message for errors.
+    # 3. validate the forms.
+
+    my $process = $c->stash->{process};
+    my %messages;
+    my @validations;
+    my $field = $process->first_field;
+
+    while ($field) {
+
+        $messages{ $field->id } =  { 
+	        DEFAULT => $field->description ? $field->description : "input ".$field->name." is invalid",
+	        ($field->type == 0 or $field->type == 2) ?
+	        ('LENGTH' => "input ".$field->name." ( length should be between ".$field->min_length." and ".$field->max_length." )") :
+	        ('BETWEEN' => "input ".$field->name." ( value should be between ".$field->min_length." and ".$field->max_length." )"),
+            }; 
+        
+        my %valid_field = (
+            $field->id => [ 
+	        'NOT_BLANK', 
+		($field->type == 0 or $field->type == 2) ?
+		    [ 
+		        'LENGTH', 
+		        $field->min_length ? $field->min_length : 0, 
+		        $field->max_length ? $field->max_length : $field->type == 2 ? 8192 : 64,
+		    ] :
+                    [
+		        'BETWEEN',
+		        $field->min_length ? $field->min_length : 0,
+			$field->max_length ? $field->max_length : $field->type == 1 ? 2147483647 : 1,
+		    ],
+		($field->type == 1 or $field->type == 3) ? 'INT' : 'ASCII',
+            ],
+	);
+
+        pop @{$valid_field{$field->id}} if ($field->type == 0 or $field->type == 2);
+	
+	shift @{$valid_field{$field->id}} if $field->optional == 1;
+
+	push @validations, %valid_field;
+	$field = $field->next_field;
+    }
+
+    my $validator = FormValidator::Simple->new;
+    $validator->set_messages({ add_request => {%messages} });
+    #$c->log->_dump([@validations]);
+    my $result = $validator->check($c->req => [@validations] );
+    if ( $result->has_error ) {
+        $c->stash(
+            form_errors => $result->messages("add_request"),
+	    fillform    => 1,
+        );
+    }
+
+    return $result;
+}
 =head2 requests
 
 List requests for the stashed L<UFL::Workflow::Schema::Process>.
